Frontend: Move import parsing tests into frontend suites
